Add TimeSeriesData proto message
Commit adds the TimeSeriesData and TimeSeriesDatapoint messages, simple
messages which contain time series data as it would be input from external or
internal sources. The TimeSeriesData message is a series name paired with
several TimeSeriesDatapoints, which are simply (timestamp, value) tuples.

Included is the `ToInternal()` method, which converts a TimeSeriesData
message into one or more InternalTimeSeriesData messages which are used to
efficiently store TimeSeriesData in cockroach.